requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
langchain==0.1.4
langchain-groq==0.0.1
langchain-community==0.0.16
sentence-transformers==2.3.1
supabase==2.3.4
duckduckgo-search==4.1.1
python-dotenv==1.0.0
httpx==0.26.0
numpy==1.26.3
torch==2.1.2
transformers==4.37.2
